+++ src/aafunc.js
@@ -85,6 +85,15 @@ addFuncDef({
   defaultParams: [5, 'avg'],
 });
 
+addFuncDef({
+  name: 'exclude',
+  category: 'Filter Series',
+  params: [
+    { name: 'pattern', type: 'string' },
+  ],
+  defaultParams: [],
+});
+
 // Options
 
 addFuncDef({